@@ -1,6 +1,6 @@
 {
   "name": "nebular-dashboard",
-  "version": "17.0.0",
+  "version": "18.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --proxy-config proxy.conf.json",
@@ -10,23 +10,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.0",
-    "@angular/cdk": "^17.1.0",
-    "@angular/common": "^17.0.0",
-    "@angular/compiler": "^17.0.0",
-    "@angular/core": "^17.0.0",
-    "@angular/forms": "^17.0.0",
-    "@angular/platform-browser": "^17.0.0",
-    "@angular/platform-browser-dynamic": "^17.0.0",
-    "@angular/router": "^17.0.0",
-    "@nebular/auth": "^13.0.0",
-    "@nebular/eva-icons": "^13.0.0",
-    "@nebular/security": "^13.0.0",
-    "@nebular/theme": "^13.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^16.0.0",
-    "@ngrx/router-store": "^17.1.1",
-    "@ngrx/store": "^17.1.1",
-    "@ngrx/store-devtools": "^17.1.1",
+    "@angular/animations": "^18.2.9",
+    "@angular/cdk": "^18.2.10",
+    "@angular/common": "^18.2.9",
+    "@angular/compiler": "^18.2.9",
+    "@angular/core": "^18.2.9",
+    "@angular/forms": "^18.2.9",
+    "@angular/platform-browser": "^18.2.9",
+    "@angular/platform-browser-dynamic": "^18.2.9",
+    "@angular/router": "^18.2.9",
+    "@nebular/auth": "^14.0.1",
+    "@nebular/eva-icons": "^14.0.1",
+    "@nebular/security": "^14.0.1",
+    "@nebular/theme": "^14.0.1",
+    "@ng-bootstrap/ng-bootstrap": "^17.0.1",
+    "@ngrx/router-store": "^18.1.0",
+    "@ngrx/store": "^18.1.0",
+    "@ngrx/store-devtools": "^18.1.0",
     "@ngx-translate/core": "^15.0.0",
     "@ngx-translate/http-loader": "^8.0.0",
     "@popperjs/core": "^2.11.8",
@@ -36,13 +36,13 @@
     "rxjs": "~7.8.0",
     "secure-web-storage": "^1.0.2",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.2"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.3",
-    "@angular/cli": "^17.0.3",
-    "@angular/compiler-cli": "^17.0.0",
-    "@angular/localize": "^17.0.0",
+    "@angular-devkit/build-angular": "^18.2.10",
+    "@angular/cli": "^18.2.10",
+    "@angular/compiler-cli": "^18.2.9",
+    "@angular/localize": "^18.2.9",
     "@schematics/angular": "^17.0.3",
     "@types/crypto-js": "^4.2.2",
     "@types/jasmine": "~5.1.0",
@@ -52,6 +52,6 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "typescript": "~5.2.2"
+    "typescript": "~5.5.4"
   }
 }
